+++ LibraryManagement.Application/ViewModels/LoanViewModel.cs
@@ -5,7 +5,7 @@
         public DateTime LoanDate { get; private init; }
         public DateTime? DevolutionDate { get; private init; }
         public UserViewModel User { get; private init; }
-        public int TotalLoanBooks => Books.Count;
+        public sbyte TotalLoanBooks => (sbyte)Books.Count;
         public List<BookViewModel> Books { get; private init; } = [];
 
         public LoanViewModel(DateTime loanDate, DateTime? devolutionDate, UserViewModel user)
